Use document.URL instead of location.href.
location.href on Safari has a bug where it cannot tell the difference between
a real # and an encoded #.  Since Safari does not support ResTiming, this isn't
an issue, but if at some point they do start supporting it, this could cause a
problem.

It turns out that document.URL is exactly the value of location.href, but w/o
the bugs.